Atualizando a funcionalidade divisao

+++ src/Operacoes/Divisao.java
@@ -1,33 +1,33 @@
 package Operacoes;
 
 public class Divisao {
-    private int num1;
-    private int num2;
+    private double num1;
+    private double num2;
 
-    public Divisao(int num1, int num2) {
+    public Divisao(double num1, double num2) {
         this.num1 = num1;
         this.num2 = num2;
     }
 
-    public int getNum1() {
+    public double getNum1() {
         return num1;
     }
 
-    public void setNum1(int num1) {
+    public void setNum1(double num1) {
         this.num1 = num1;
     }
 
-    public int getNum2() {
+    public double getNum2() {
         return num2;
     }
 
-    public void setNum2(int num2) {
+    public void setNum2(double num2) {
         this.num2 = num2;
     }
 
-    public void operacaoDivisao(int num1, int num2) {
+    public void operacaoDivisao(double num1, double num2) {
         try {
-            int resultado = num1 / num2;
+            double resultado = num1 / num2;
             System.out.println(+getNum1()+ " / "+ getNum2()+ " = "+resultado);
             
         } catch (ArithmeticException e) {
